Lint examples
Jsx only in .jsx files
